[1632567] Greedy explore now tries to do the Right Thing with mimics.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@915 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/stuff.h
@@ -102,7 +102,7 @@ int letter_to_index(int the_letter);
 
 int near_stairs(int px, int py, int max_dist, unsigned char &stair_gfx);
 
-inline bool testbits(unsigned int flags, unsigned int test)
+inline bool testbits(unsigned long flags, unsigned long test)
 {
     return ((flags & test) == test);
 }
